requirements.txt
streamlit
groq
load_dotenv
pydub
gtts
numpy
pydub
SpeechRecognition
pyaudio
pygame
deepgram-sdk